+++ MachoLoader/Classes/SymbolicInfo.h
@@ -12,18 +12,18 @@
 
 	NSString	*_segmentName;
 	NSString	*_sectionName;
-	NSUInteger	_address;
+	uint64		_address;
 	NSString	*_stringValue;
 	NSString	*_libraryName;
-	long long	_intValue;
+	uint64		_intValue;
 	double		_floatValue;
 }
 
 @property (retain) NSString		*segmentName;
 @property (retain) NSString		*sectionName;
-@property (assign) NSUInteger	address;
+@property (assign) uint64		address;
 @property (retain) NSString		*stringValue;
 @property (retain) NSString		*libraryName;
-@property (assign) long long	intValue;
+@property (assign) uint64		intValue;
 @property (assign) double		floatValue;
 @end
